+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - language option in github api
 - async utils
+- download to file with headers and cookies options
 
 ### Removed
 - print_item_info() in time.profile